fix(daemon): store unit tests (#2700)
* fix(daemon): fix peapod unit test

* fix(daemon): fix workspace store test

* fix(daemon): fix flow store unit test

* fix(daemon): fix more tests

* fix(daemon): fix test logs

* fix(daemon): fix more tests

* fix(daemon): unify workspace fixture

* fix(daemon): increase test timeout to 30 min

* fix(daemon): remove docker host for flow creation

* fix(daemon): set ci timeout back to 20

* fix(daemon): fix test files

+++ daemon/stores/workspaces.py
@@ -55,19 +55,20 @@ class WorkspaceStore(BaseStore):
         return id
 
     def rm_files(self, id: DaemonID, logs: bool = False):
-        workdir = self[id].metadata.workdir
-        if not workdir or not Path(workdir).is_dir():
-            self._logger.info(f'there\'s nothing to remove in workdir {workdir}')
-            return
-        if logs:
-            self._logger.info(f'asked to remove complete directory: {workdir}')
-            rmtree(workdir)
-            self[id].metadata.workdir = ''
-        else:
-            for path in Path(workdir).rglob('[!logging.log]*'):
-                if path.is_file():
-                    self._logger.debug(f'file to be deleted: {path}')
-                    path.unlink()
+        if self[id].metadata:
+            workdir = self[id].metadata.workdir
+            if not workdir or not Path(workdir).is_dir():
+                self._logger.info(f'there\'s nothing to remove in workdir {workdir}')
+                return
+            if logs:
+                self._logger.info(f'asked to remove complete directory: {workdir}')
+                rmtree(workdir)
+                self[id].metadata.workdir = ''
+            else:
+                for path in Path(workdir).rglob('[!logging.log]*'):
+                    if path.is_file():
+                        self._logger.debug(f'file to be deleted: {path}')
+                        path.unlink()
 
     def rm_network(self, id: DaemonID):
         try:
@@ -132,7 +133,7 @@ class WorkspaceStore(BaseStore):
             raise KeyError(f'{colored(str(id), "cyan")} not found in store.')
 
         # Peas/Pods/Flows need to be deleted before networks, files etc can be deleted
-        if everything:
+        if everything and self[id].metadata:
             from . import get_store_from_id
 
             ids_to_delete = list(self[id].metadata.managed_objects)
